refactor: connect calendar to store
